Move FormActivity and FormContract to form package
