+++ paymentsheet/src/main/AndroidManifest.xml
@@ -38,7 +38,7 @@
             android:name=".paymentdatacollection.cvcrecollection.CvcRecollectionActivity"
             android:theme="@style/StripePaymentSheetDefaultTheme" />
         <activity
-            android:name="com.stripe.android.paymentelement.embedded.FormActivity"
+            android:name="com.stripe.android.paymentelement.embedded.form.FormActivity"
             android:theme="@style/StripePaymentSheetDefaultTheme" />
         <activity
             android:name="com.stripe.android.paymentelement.embedded.manage.ManageActivity"
